% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china_map}
\alias{china_map}
\title{China Map Data}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{province}{name of province/municipality/autonomous region/SAR}
  \item{long}{longitude}
  \item{lat}{latitude}
  \item{group}{groups of individual polygons; used in plotting, e.g. ggplot(china_map, aes(long, lat, group=group))+geom_polygon()}
  ...
}}
\source{
\url{http://bzdt.ch.mnr.gov.cn}
}
\usage{
china_map
}
\description{
China map data with boundaries between provinces/municipalities/
autonomous regions/SARs
}
\keyword{datasets}
