% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china_map_load.R
\name{china_map_load}
\alias{china_map_load}
\title{Translate province}
\usage{
china_map_load(decimal = NULL)
}
\arguments{
\item{decimal}{number of decimals kept. By default it's kept to the best precision.}
}
\value{
the map data as a tibble
}
\description{
Translate province
}
\examples{
china_map_load(2)
}
