% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{province_data}
\alias{province_data}
\title{Province Data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 34 rows and 6 columns.}
\usage{
province_data
}
\description{
Data of provinces/municipalities/autonomous regions/SARs,
including the positions of their capitals
}
\keyword{datasets}
