% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{province_translation}
\alias{province_translation}
\title{Translation of province names}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{...}{short or full names for each language}
  ...
}}
\usage{
province_translation
}
\description{
Translations of provinces/municipalities/autonomous regions/SARs
}
\keyword{datasets}
