% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_province}
\alias{translate_province}
\title{Translate province}
\usage{
translate_province(data, to = NULL)
}
\arguments{
\item{data}{A tibble/df with `province` column containing full Chinese names of provinces `china_map` table}

\item{to}{`cn_short`, `en_full`, `en_short`, `ja_long`, etc. If not specified, all translations will be joined.}
}
\value{

}
\description{
Translate province
}
